+++ src/apps/external_executor/external_executor_indexing.h
@@ -15,11 +15,27 @@
 
 namespace externalexecutor
 {
+  class MapIndex;
+
+  using MapStrategyPtr = std::shared_ptr<MapIndex>;
   using DetachedIndexStream =
     ccf::grpc::DetachedStreamPtr<externalexecutor::protobuf::IndexWork>;
   using IndexStream =
     ccf::grpc::StreamPtr<externalexecutor::protobuf::IndexWork>;
   using BucketValue = std::pair<ccf::indexing::FetchResultPtr, std::string>;
+
+  class ImplMapIndex
+  {
+    LRU<std::string, std::string> indexed_data;
+    std::unordered_map<std::string, BucketValue> results_in_progress;
+    std::shared_ptr<ccf::indexing::AbstractLFSAccess> lfs_access;
+
+  public:
+    ImplMapIndex(
+      const std::shared_ptr<ccf::indexing::AbstractLFSAccess>& lfs_access_);
+    std::optional<std::string> fetch_data(std::string& key);
+    void store_data(std::string& key, std::string& value);
+  };
 
   class MapIndex : public ccf::indexing::Strategy
   {
@@ -29,175 +45,135 @@ namespace externalexecutor
     ccf::TxID current_txid = {};
     ExecutorId indexer_id;
     ccf::endpoints::CommandEndpointContext* endpoint_ctx;
+    ccfapp::AbstractNodeContext& node_context;
     IndexStream out_stream;
     bool is_indexer_active = false;
     DetachedIndexStream detached_stream;
+    std::shared_ptr<ImplMapIndex> impl_index = nullptr;
 
   public:
-    // store unserialized indexed Key-Value data
-    // wrap it in LRU?
-    LRU<std::string, std::string> indexed_data_;
-    std::unordered_map<std::string, std::string> indexed_data;
-    std::unordered_map<std::string, BucketValue> results_in_progress;
-
     MapIndex(
       const std::string& map_name_,
       const std::string& strategy_prefix,
       ExecutorId& id,
       ccf::endpoints::CommandEndpointContext& ctx,
-      IndexStream&& stream) :
-      Strategy(strategy_prefix),
-      map_name(map_name_),
-      indexer_id(id),
-      endpoint_ctx(&ctx),
-      out_stream(std::move(stream)),
-      is_indexer_active(true),
-      indexed_data_(10)
-    {
-      if (kv::get_security_domain(map_name_) != kv::SecurityDomain::PUBLIC)
-      {
-        throw std::logic_error(fmt::format(
-          "This Strategy ({}) is currently only implemented for public tables, "
-          "so cannot be used for '{}'",
-          get_name(),
-          map_name_));
-      }
-      // create a detached stream pointer of the indexer
-      detached_stream =
-        ccf::grpc::detach_stream(ctx.rpc_ctx, std::move(out_stream), [this]() {
-          is_indexer_active = false;
-        });
-    }
+      ccfapp::AbstractNodeContext& node_context,
+      IndexStream&& stream);
 
     void handle_committed_transaction(
-      const ccf::TxID& tx_id, const kv::ReadOnlyStorePtr& store) override
-    {
-      auto tx = store->create_read_only_tx();
-      auto handle = tx.ro<Map>(map_name);
+      const ccf::TxID& tx_id, const kv::ReadOnlyStorePtr& store) override;
 
-      handle->foreach([this](const auto& k, const auto& v) {
-        externalexecutor::protobuf::IndexWork data;
-        externalexecutor::protobuf::IndexKeyValue* index_key_value =
-          data.mutable_key_value();
-        index_key_value->set_key(k);
-        index_key_value->set_value(v);
-        if (is_indexer_active)
-        {
-          // stream transactions to the indexer
-          if (!out_stream->stream_msg(data))
-          {
-            LOG_DEBUG_FMT("Failed to stream request to indexer {}", indexer_id);
-          }
-        }
+    std::optional<ccf::SeqNo> next_requested() override;
 
-        return true;
-      });
-      current_txid = tx_id;
-    }
+    void store(std::string& key, std::string& value);
 
-    std::optional<ccf::SeqNo> next_requested() override
-    {
-      return current_txid.seqno + 1;
-    }
+    std::optional<std::string> fetch(std::string& key);
   };
 
-  // Note: move this into a different scope
-  using MapStrategyPtr = std::shared_ptr<MapIndex>;
-  std::unordered_map<std::string, MapStrategyPtr> map_index_strategies;
-  std::shared_ptr<ccf::indexing::AbstractLFSAccess> lfs_access;
-
-  static std::optional<std::string> fetch_data(
-    MapStrategyPtr strategy, std::string& key)
+  class LRUIndex
   {
-    bool complete = true;
-    ccf::pal::Mutex results_access;
+  public:
+    using Entry = std::pair<const std::string, std::string>;
+    using List = std::list<Entry>;
+    using Map = std::map<std::string, typename List::iterator>;
+    using Iterator = typename List::iterator;
+    using ConstIterator = typename List::const_iterator;
 
-    while (true)
+  private:
+    // Entries are ordered by when they were most recently accessed, with most
+    // recent at the front
+    List entries_list;
+
+    // Maps from keys to iterators from entries_list, which must remain valid
+    // even when entries_list is modified
+    Map iter_map;
+
+    std::shared_ptr<ccf::indexing::AbstractLFSAccess> lfs_access;
+    size_t max_size;
+
+    void cull()
     {
-      std::lock_guard<ccf::pal::Mutex> guard(results_access);
-
-      const auto old_it = strategy->results_in_progress.find(key);
-      if (old_it != strategy->results_in_progress.end())
+      while (entries_list.size() > max_size)
       {
-        auto& bucket_value = old_it->second;
+        const auto& least_recent_entry = entries_list.back();
+        auto key = least_recent_entry.first;
+        auto value = least_recent_entry.second;
+        auto hex_key = ds::to_hex(key.begin(), key.end());
+        std::string blob_key = fmt::format("{}:{}", "table-name", hex_key);
+        ccf::indexing::LFSContents contents(value.begin(), value.end());
+        lfs_access->store(blob_key, std::move(contents));
+        iter_map.erase(least_recent_entry.first);
+        entries_list.pop_back();
+      }
+    }
 
-        // We were already trying to fetch this. If it's finished fetching,
-        // parse and store the result
-        if (bucket_value.first != nullptr)
-        {
-          const auto fetch_result = bucket_value.first->fetch_result;
-          switch (fetch_result)
-          {
-            case (ccf::indexing::FetchResult::Fetching):
-            {
-              complete = false;
-              break;
-            }
-            case (ccf::indexing::FetchResult::Loaded):
-            {
-              bool corrupt = false;
-              std::string val(
-                bucket_value.first->contents.begin(),
-                bucket_value.first->contents.end());
-              bucket_value.second = val;
-              if (!corrupt)
-              {
-                bucket_value.first = nullptr;
-                break;
-              }
-              else
-              {
-                // Deliberately fall through to the case below. If this can't
-                // deserialise the value, consider the file corrupted
-                LOG_FAIL_FMT("Deserialisation failed");
-              }
-            }
-            case (ccf::indexing::FetchResult::NotFound):
-            case (ccf::indexing::FetchResult::Corrupt):
-            {
-              complete = false;
-              const auto problem =
-                fetch_result == ccf::indexing::FetchResult::NotFound ?
-                "missing" :
-                "corrupt";
-              LOG_FAIL_FMT(
-                "A file that indexer requires is {}. Re-indexing.", problem);
-              LOG_DEBUG_FMT(
-                "The {} file is {}", problem, bucket_value.first->key);
+  public:
+    LRUIndex(
+      std::shared_ptr<ccf::indexing::AbstractLFSAccess> lfs_ptr,
+      size_t max_size) :
+      lfs_access(lfs_ptr),
+      max_size(max_size)
+    {}
 
-              strategy->results_in_progress.clear();
-              strategy->indexed_data.clear();
-              return std::nullopt;
-              break;
-            }
-          }
-        }
+    size_t size() const
+    {
+      return iter_map.size();
+    }
+
+    void set_max_size(size_t ms)
+    {
+      max_size = ms;
+      cull();
+    }
+
+    size_t get_max_size() const
+    {
+      return max_size;
+    }
+
+    std::optional<std::string> find(const std::string& k)
+    {
+      const auto it = iter_map.find(k);
+      if (it != iter_map.end())
+      {
+        return it->second->second;
+      }
+
+      return std::nullopt;
+    }
+
+    Iterator insert(const std::string& k, std::string&& v)
+    {
+      auto it = iter_map.find(k);
+      if (it != iter_map.end())
+      {
+        // If it already exists, move to the front
+        auto& list_it = it->second;
+        entries_list.splice(entries_list.begin(), entries_list, list_it);
       }
       else
       {
-        // We're not currently fetching this. First check if it's in our
-        // current results
-        const auto current_it = strategy->indexed_data.find(key);
-        if (current_it != strategy->indexed_data.end())
-        {
-          if (complete)
-          {
-            return strategy->indexed_data[key];
-          }
-        }
-
-        else
-        {
-          // Begin fetching this bucket from disk
-          auto hex_key = ds::to_hex(key.begin(), key.end());
-          std::string blob_key = fmt::format("{}:{}", "table-name", hex_key);
-          auto fetch_handle = lfs_access->fetch(blob_key);
-          std::string value;
-          strategy->results_in_progress[key] =
-            std::make_pair(fetch_handle, value);
-          complete = false;
-        }
+        // Else add a new entry to both containers, and cull if necessary
+        entries_list.push_front(
+          std::make_pair(k, std::forward<std::string>(v)));
+        const auto list_it = entries_list.begin();
+        iter_map.emplace_hint(it, k, list_it);
+        cull();
       }
+
+      return entries_list.begin();
     }
-  }
+
+    std::string& operator[](std::string&& k)
+    {
+      auto it = insert(std::forward<std::string>(k), {});
+      return it->second;
+    }
+
+    void clear()
+    {
+      entries_list.clear();
+      iter_map.clear();
+    }
+  };
 }
